backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
anthropic==0.34.0
pydantic==2.5.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
requests==2.31.0
